Migrate all servlets to membership servlet. Polish the features from last hw2.
