+++ web/home.jsp
@@ -7,5 +7,7 @@
 <%@page contentType="text/html" pageEncoding="UTF-8"%>
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
 <c:import url="/includes/header.jsp" />
+
 <h1>Home JSP</h1>
+
 <c:import url="/includes/footer.jsp" />
