+++ web/login.jsp
@@ -13,13 +13,12 @@
 <c:if test="${message != null}">
     <span>${message}</span>
 </c:if>
-<form action="login" method="post">
+<form action="membership" method="post">
     <input type="hidden" name="action" value="authenticate">
     <label class="pad_top">Email:</label>
     <input type="email" name="email" value="${user.email}"><br>
     <label class="pad_top">Password:</label>
-    <input type="text" name="password" value="${user.password}"><br>
-
+    <input type="password" name="password" value="${user.password}"><br>
     <input type="submit" value="Log in" class="margin_left">
     <a href='./forgotpassword.jsp'>Forgot password?</a>
     
@@ -27,5 +26,4 @@
         New? <a href='./signup.jsp'>Sign up now!</a>
     </p>
 </form>
-    </body>
 <c:import url="/includes/footer.jsp" />
